% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_calcPolyfunctionality.R
\name{calcPolyfunctionality}
\alias{calcPolyfunctionality}
\title{Calculate the frequency of polyfunctional cells}
\usage{
calcPolyfunctionality(
  x,
  md_cols,
  pop_col = "cell_type",
  resolution = "MarkerComb",
  condition_col = "Stimulation",
  background_val = "unstimulated"
)
}
\arguments{
\item{x}{A dataframe generated by the \code{\link{assignMarkerCombinations}} function,
containing marker combination data along with associated metadata.}

\item{md_cols}{A character vector specifying the names of metadata columns in \code{x}.
These columns provide contextual information (e.g., participant ID, sample ID,
time point).}

\item{pop_col}{A string giving the name of the column in \code{x} that contains the
population of interest. The function calculates the percentage of positive cells for
each marker combination within this population. Can be obtained from any source, e.g.,
annotated metacluster IDs from FlowSOM clustering or cell labels from manual gating
(see \code{\link[FlowSOM]{GetFlowJoLabels}}).}

\item{resolution}{A string specifying the column in \code{x} that defines
polyfunctionality, e.g., \code{"MarkerComb"}, \code{"MarkerCount"}, or
\code{"Functionality"}. Default is \code{"MarkerComb"}. Can also refer to any of the
 logical columns generated if \code{mode = "exhaustive"} was used in
\code{\link{assignMarkerCombinations}}, or to custom resolutions created by combining
outputs from \code{\link{assignMarkerCombinations}}. For example,
\code{atleast_IFN+TNFa+ == TRUE & MarkerCount == 3} refers to cells with three
functions positive for at least IFN and TNFa. Users must first create the logical
column in \code{x} (e.g., using \code{dplyr::mutate()}) before specifying it here.}

\item{condition_col}{A string giving the name of the column in \code{x} containing the
stimulation condition (e.g., antigen-stimulated, unstimulated). Default =
\code{"Stimulation"}.}

\item{background_val}{A value from \code{condition_col} indicating the background
condition (e.g., "unstimulated"). This is used as the reference for stimulation
conditions. Default = \code{"unstimulated"}.}
}
\value{
A dataframe similar to \code{x}, with the addition of four new columns:
  \itemize{
    \item{\code{n_cells_pos}: Number of positive cells}
    \item{\code{n_cells_tot}: Number of total cells}
    \item{\code{perc_pos}: Percentage positive cells, i.e.
      \code{(n_cells_pos/n_cells_tot)*100}}
    \item{\code{perc_pos_backgroundSubtracted}: Percentage positive cells with
      corresponding background subtracted. Negative values are set to zero.}
  }
}
\description{
Calculate the frequency of cells positive for each marker combination within a defined
population and subtract the corresponding background. Negative values are set to zero.
}
\examples{
# Define path to data
path_data <- system.file("extdata", package = "polyICSFlow")

# Read fcs files as flowSet
fs <- flowCore::read.flowSet(path = path_data,
                             pattern = ".fcs",
                             truncate_max_range = FALSE,
                             transformation = FALSE)

# Create GatingSet
gs <- flowWorkspace::GatingSet(fs)

# Apply GatingTemplate
gt <- openCyto::gatingTemplate(file.path(path_data,
                                        "gatingTemp_cytokines.csv"))
openCyto::gt_gating(x = gt, y = gs)

# Get marker positivity per cell
df_markerPos <- getMarkerPositivity(x = gs,
                                    gate_names = c("IFN+","TNFa+","IL2+","CD107a+"))

# Assign marker combinations to each cell based on marker positivity
df_markerComb <- assignMarkerCombinations(data_markerPos = df_markerPos,
                                          mode = "exhaustive")

# Read metadata and cell labels
df_md <- read.csv(file.path(path_data, "metadata.csv"))
metacluster_labels <- readRDS(file.path(path_data, "cell_labels.rds"))

# Add metadata and cell labels to dataframe
df <- df_markerComb \%>\%
    dplyr::mutate(ID = df_md$ID,
                  Stimulation = df_md$Stimulation,
                  cell_type = metacluster_labels)

# Calculate by Marker Combination
result <- calcPolyfunctionality(x = df,
                                md_cols = c("ID"),
                                pop_col = "cell_type",
                                resolution = "MarkerComb",
                                condition_col = "Stimulation",
                                background_val = "Unstim")

# Calculate by Marker Count
result <- calcPolyfunctionality(x = df,
                                md_cols = c("ID"),
                                pop_col = "cell_type",
                                resolution = "MarkerCount",
                                condition_col = "Stimulation",
                                background_val = "Unstim")

# Calculate by Functionality
result <- calcPolyfunctionality(x = df,
                                md_cols = c("ID"),
                                pop_col = "cell_type",
                                resolution = "Functionality",
                                condition_col = "Stimulation",
                                background_val = "Unstim")

# Calculate by logical at least n
result <- calcPolyfunctionality(x = df,
                                md_cols = c("ID"),
                                pop_col = "cell_type",
                                resolution = "atleast_2_IFN+TNFa+",
                                condition_col = "Stimulation",
                                background_val = "Unstim")

# Calculate by logical Polyfunctional, at least marker
result <- calcPolyfunctionality(x = df,
                                md_cols = c("ID"),
                                pop_col = "cell_type",
                                resolution = "Polyfunctional_atleast_IFN+",
                                condition_col = "Stimulation",
                                background_val = "Unstim")

# Calculate by custom resolution
df2 <- df \%>\%
  dplyr::mutate(MarkerCountAbove3_atleast_IFN_TNFa = MarkerCount == 3 &
                 .data[["atleast_2_IFN+TNFa+"]])

result <- calcPolyfunctionality(x = df2,
                                md_cols = c("ID"),
                                pop_col = "cell_type",
                                resolution = "MarkerCountAbove3_atleast_IFN_TNFa",
                                condition_col = "Stimulation",
                                background_val = "Unstim")


}
\seealso{
\code{\link{getMarkerPositivity}}, \code{\link{assignMarkerCombinations}}
}
