% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extractCellCounts.R
\name{extractCellCounts}
\alias{extractCellCounts}
\title{Get number of cells positive for each distinct marker combination}
\usage{
extractCellCounts(
  data_markerComb,
  cell_subset_params = list(cellTypes = NULL, cellTypes_plot = NULL)
)
}
\arguments{
\item{data_markerComb}{A dataframe generated by the \code{\link{assignMarkerCombinations}} function,
containing marker combination data.}

\item{cell_subset_params}{List with parameters describing what subset of cells (cellTypes_plot) should be counted,
from a vector of cell types (cellTypes). Default = \code{list(cellTypes = NULL, cellTypes_plot = NULL)},
i.e. all cells in the dataset are included in the cell count.}
}
\value{
A grouped tibble with cell counts of the input data grouped by \code{"MarkerComb"}, \code{"MarkerCount"}, and
  \code{"Functionality"}.
}
\description{
Get number of cells in your dataset positive for each distinct marker combination (MarkerComb),
 number of functions (MarkerCount), and Functionality.
}
\examples{
# Define path to data
path_data <- system.file("extdata", package = "polyICSFlow")

# Read fcs files as flowSet
fs <- flowCore::read.flowSet(path = path_data,
                             pattern = ".fcs",
                             truncate_max_range = FALSE,
                             transformation = FALSE)

# Create GatingSet
gs <- flowWorkspace::GatingSet(fs)

# Apply GatingTemplate
gt <- openCyto::gatingTemplate(file.path(path_data, "gatingTemp_cytokines.csv"))
openCyto::gt_gating(x = gt, y = gs)

# Get marker positivity per cell
df_markerPos <- getMarkerPositivity(x = gs,
                                    gate_names = c("IFN+","TNFa+","IL2+","CD107a+"))

# Assign marker combinations to each cell based on marker positivity
df_markerComb <- assignMarkerCombinations(data_markerPos = df_markerPos,
                                          mode = "exhaustive")

# Get cell counts from all data
cell_counts <- extractCellCounts(data_markerComb = df_markerComb)

# Get cell counts from subset of data (metacluster 2 and 4)
metacluster_labels <- readRDS(file.path(path_data, "cell_labels.rds"))
cell_counts <- extractCellCounts(data_markerComb = df_markerComb,
                                 cell_subset_params = list(cellTypes = metacluster_labels,
                                                           cellTypes_plot = c(2,4)))
}
