% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_getMarkerPositivity.R
\name{getMarkerPositivity}
\alias{getMarkerPositivity}
\title{Create dataframe of marker positivity}
\usage{
getMarkerPositivity(x, gate_names = NULL)
}
\arguments{
\item{x}{An object of class \code{GatingHierarchy} or \code{GatingSet} with an applied \code{gatingTemplate}.}

\item{gate_names}{The names of the cytokine gates. If  \code{NULL} (default), all leaf nodes will be used.}
}
\value{
A data.frame where each row is a cell and each column is a gated marker which can be used as input for
the \code{\link{assignMarkerCombinations}} function.
}
\description{
Create a dataframe describing whether each cell is included by the gates where each row is a cell and each column is a marker.
}
\examples{
# Define path to data
path_data <- system.file("extdata", package = "polyICSFlow")

# Read fcs files as flowSet
fs <- flowCore::read.flowSet(path = path_data,
                             pattern = ".fcs",
                             truncate_max_range = FALSE,
                             transformation = FALSE)

# Create GatingSet
gs <- flowWorkspace::GatingSet(fs)

# Apply GatingTemplate
gt <- openCyto::gatingTemplate(file.path(path_data, "gatingTemp_cytokines.csv"))
openCyto::gt_gating(x = gt, y = gs)

# Get marker positivity per cell
df_markerPos <- getMarkerPositivity(x = gs,
                                    gate_names = c("IFN+","TNFa+","IL2+","CD107a+"))
}
\references{
This code is strongly based on the \code{\link[FlowSOM]{GetFlowJoLabels}} function.
}
\seealso{
\code{\link[flowWorkspace]{GatingSet}}, \code{\link[openCyto]{gatingTemplate}},
\code{\link[openCyto]{gt_gating}}, \code{\link{assignMarkerCombinations}},
\code{\link{calcPolyfunctionality}}
}
