% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{plotMarkerCombHeatmap}
\alias{plotMarkerCombHeatmap}
\title{Plot a heatmap of all possible marker combinations.}
\usage{
plotMarkerCombHeatmap(
  marker_names,
  orientation = c("vertical", "horizontal"),
  palette_count = NULL,
  palette_comb = NULL,
  full_names = TRUE
)
}
\arguments{
\item{marker_names}{A character vector specifying the marker names used to
generate all possible combinations.}

\item{orientation}{A string specifying the orientation of the heatmap, either vertical or horizontal.
Default = \code{"vertical"}.}

\item{palette_count}{A named character vector mapping the number of markers
per cell to colors. If \code{NULL} (default), a qualitative palette is generated with
\code{\link[scales]{brewer_pal}}.}

\item{palette_comb}{A named character vector mapping marker combination
names to colors. If \code{NULL} (default), a qualitative palette is generated with
\code{CATALYST:::.cluster_cols}.}

\item{full_names}{Logical. If TRUE (default) the marker combinations are displayed with their full
names (i.e., negative markers are also present). If FALSE, only the positive markers are displayed.}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap}} object visualizing all marker
  combinations. The heatmap includes two layers of annotation:
  \itemize{
    \item \code{MarkerCount}: A marker count annotation (colored
      with \code{palette_count}) showing the number of markers in each combination.
    \item \code{MarkerComb}: A combination annotation (colored
      with \code{palette_comb}) showing the exact subset of markers present.
  }
}
\description{
Plot a heatmap of all possible marker combinations.
}
\examples{
# Vertical heatmap
plotMarkerCombHeatmap(marker_names = c("IFN","TNFa","IL2","CD107a"),
                      orientation = "vertical")

# Horizontal heatmap
plotMarkerCombHeatmap(marker_names = c("IFN","TNFa","IL2","CD107a"),
                      orientation = "horizontal")
}
\seealso{
\code{\link[ComplexHeatmap]{Heatmap}}
}
