% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_assignMarkerCombinations.R
\name{assignMarkerCombinations}
\alias{assignMarkerCombinations}
\title{Assing each cell to a combination of gated markers.}
\usage{
assignMarkerCombinations(data_markerPos, mode = c("simple", "exhaustive"))
}
\arguments{
\item{data_markerPos}{A data.frame where each row is a cell and each column is a gated marker describing
whether a cell is included in the gate, as constructed by the \code{\link{getMarkerPositivity}} function.}

\item{mode}{Character string specifying the level of detail in the polyfunctionality assignment.
If \code{simple}, the output includes only \code{MarkerComb} (exact marker combinations),
\code{MarkerCount} (number of functions), and \code{Functionality} (mono/polyfunctional).
If \code{exhaustive}, the output additionally contains all "atleast_n" columns, which are logical
describing whether cells have at least 1, 2, 3, ... \eqn{n} markers in any combination. Default = \code{simple}.}
}
\value{
The input data.frame with three extra columns describing
    \itemize{
      \item{\code{MarkerComb}: A factor describing which marker combination a specific cell has. Describes
      the exact combinations, which are mutually exclusive (e.g., 4 markers → 2⁴ = 16 combinations).}
      \item{\code{MarkerCount}: An integer describing the number of functions of the cell (count of positive markers).}
      \item{\code{Functionality}: A factor grouping \code{MarkerCount)} into three categories; No cytokines (\code{MarkerCount == 0}),
      Monofunctional (\code{MarkerCount == 1}), or Polyfunctional (\code{MarkerCount > 1}).}
    }
}
\description{
Assing each cell to a combination of gated markers.
}
\examples{
# Define path to data
path_data <- system.file("extdata", package = "polyICSFlow")

# Read fcs files as flowSet
fs <- flowCore::read.flowSet(path = path_data,
                             pattern = ".fcs",
                             truncate_max_range = FALSE,
                             transformation = FALSE)

# Create GatingSet
gs <- flowWorkspace::GatingSet(fs)

# Apply GatingTemplate
gt <- openCyto::gatingTemplate(file.path(path_data, "gatingTemp_cytokines.csv"))
openCyto::gt_gating(x = gt, y = gs)

# Get marker positivity per cell
df_markerPos <- getMarkerPositivity(x = gs,
                                    gate_names = c("IFN+","TNFa+","IL2+","CD107a+"))

# Assign marker combinations to each cell based on marker positivity
# Simple mode
df_markerComb <- assignMarkerCombinations(data_markerPos = df_markerPos,
                                          mode = "simple")
# Exhaustive mode
df_markerComb <- assignMarkerCombinations(data_markerPos = df_markerPos,
                                          mode = "exhaustive")
}
\seealso{
\code{\link{getMarkerPositivity}}, \code{\link{calcPolyfunctionality}}
}
