% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_plotFunctions.R
\name{plotMarkerComb2DScatters}
\alias{plotMarkerComb2DScatters}
\title{Plot 2D scatterplots of all marker combinations.}
\usage{
plotMarkerComb2DScatters(
  input,
  assigned_MarkerCombs,
  seed = 1,
  cell_subset_params = list(cellTypes = NULL, cellTypes_plot = NULL),
  palette_comb = NULL,
  return_list = FALSE
)
}
\arguments{
\item{input}{Either a matrix of marker expression or an object of class
\code{\link[SingleCellExperiment]{SingleCellExperiment}}, \code{\link[flowWorkspace]{GatingSet}},
 or \code{\link[flowCore]{flowSet}}, from which the expression data can be extracted.
 If input is either GatingSet or flowSet, an aggregate of the data is created with the
 \code{\link[FlowSOM]{AggregateFlowFrames}} function.}

\item{assigned_MarkerCombs}{A vector of assigned marker combinations per
cell in \code{x}, as created by \code{\link{assignMarkerCombinations}}.}

\item{seed}{Set a seed for reproducible results (point jitters and potential downsampling). Default = 1.}

\item{cell_subset_params}{List with parameters describing what subset of cells (cellTypes_plot) should be plotted,
from a vector of cell types (cellTypes). Default = \code{list(cellTypes = NULL, cellTypes_plot = NULL)},
i.e. all cells in the dataset are plotted.}

\item{palette_comb}{A named character vector mapping marker combination
names (as produced by \code{\link{assignMarkerCombinations}}) to colors.
If \code{NULL} (default), a qualitative palette is generated with \code{\link[grDevices:colorRampPalette]{colorRampPalette}}
and \code{CATALYST:::.cluster_cols}.}

\item{return_list}{Logical. If FALSE (default), the plots are combined by cowplot. If TRUE, a list of
plots are returned to allow further customization.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object visualizing 2D scatterplots of all
  marker combinations.
}
\description{
Plots 2D scatter plots of all possible marker combinations, where each panel (facet) corresponds to one of the
\eqn{2^n} possible combinations of \eqn{n} markers.
The top-left panel represents cells negative for all markers, while the bottom-right represents cells positive for all markers,
with intermediate panels showing mixed combinations.
Within each panel, the x-axis denotes the \eqn{n} markers and the y-axis shows the transformed MFI values.
Each point corresponds to a single cell, with cells belonging to the combination represented in that facet
highlighted in color, and all other cells displayed in grey background for reference.
}
\examples{
# Define path to data
path_data <- system.file("extdata", package = "polyICSFlow")

# Read fcs files as flowSet
fs <- flowCore::read.flowSet(path = path_data,
                             pattern = ".fcs",
                             truncate_max_range = FALSE,
                             transformation = FALSE)

# Create GatingSet
gs <- flowWorkspace::GatingSet(fs)

# Apply GatingTemplate
gt <- openCyto::gatingTemplate(file.path(path_data, "gatingTemp_cytokines.csv"))
openCyto::gt_gating(x = gt, y = gs)

# Get marker positivity per cell
df_markerPos <- getMarkerPositivity(x = gs,
                                    gate_names = c("IFN+","TNFa+","IL2+","CD107a+"))

# Assign marker combinations to each cell based on marker positivity
df_markerComb <- assignMarkerCombinations(data_markerPos = df_markerPos)

# Plot 2D scatter of marker combinations in dataset (all data)
# From gatingSet
plotMarkerComb2DScatters(input = gs,
                         assigned_MarkerCombs = df_markerComb$MarkerComb)

# From flowSet
plotMarkerComb2DScatters(input = fs,
                         assigned_MarkerCombs = df_markerComb$MarkerComb)

# Plot only subset of data (metacluster 2 and 4)
metacluster_labels <- readRDS(file.path(path_data, "cell_labels.rds"))
plotMarkerComb2DScatters(input = fs,
                         assigned_MarkerCombs = df_markerComb$MarkerComb,
                         cell_subset_params = list(cellTypes = metacluster_labels,
                                                   cellTypes_plot = c(2,4)))
# As list
plot_list <- plotMarkerComb2DScatters(input = fs,
                                      assigned_MarkerCombs = df_markerComb$MarkerComb,
                                      return_list = TRUE)
}
